Missing verification on delete

+++ platform/trewny/models/Bookmark.php
@@ -35,8 +35,9 @@ class Bookmark extends \common\models\db\Bookmark
 
     public function afterDelete() {
         $name = $this->image;
-
-        unlink(Yii::$app->params['data'] . DIRECTORY_SEPARATOR . 'bookmarks' . DIRECTORY_SEPARATOR . $name);
+        if(file_exists($name)) {
+            unlink(Yii::$app->params['data'] . DIRECTORY_SEPARATOR . 'bookmarks' . DIRECTORY_SEPARATOR . $name);
+        }
     }
 
 }
